Move Spark3 Action tests to the Correct Package
Previously these files lived in o.a.iceberg.spark.actions when they
should be in o.a.iceberg.actions
