+++ spark3/src/test/java/org/apache/iceberg/actions/TestRewriteManifestsAction3.java
@@ -17,9 +17,10 @@
  * under the License.
  */
 
-package org.apache.iceberg.spark.actions;
+package org.apache.iceberg.actions;
 
-import org.apache.iceberg.actions.TestRewriteDataFilesAction;
-
-public class TestRewriteDataFilesAction3 extends TestRewriteDataFilesAction {
+public class TestRewriteManifestsAction3 extends TestRewriteManifestsAction {
+  public TestRewriteManifestsAction3(String snapshotIdInheritanceEnabled) {
+    super(snapshotIdInheritanceEnabled);
+  }
 }
